# contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: DeepBDE API
  version: 1.0.0
  description: "API for Bond Dissociation Energy (BDE) prediction using DeepBDE machine\
    \ learning model.\nBased on research from https://github.com/MSRG/DeepBDE/.\n\n\
    ## API Workflow (English)\n\n### Step 1: Get Molecule Information\n**POST** `/api/v1/predict/info/`\
    \ - Get molecule 2D structure, atoms, and bonds information\n- Input: SMILES string\n\
    - Output: Canonical SMILES, SVG image, atoms/bonds coordinates, molecule ID\n\n\
    ### Step 2: Get Canonical SMILES (Optional)\n**POST** `/api/v1/predict/info-smile-canonical/`\
    \ - Get canonical SMILES and molecule ID\n- Input: SMILES string\n- Output: Canonical\
    \ SMILES with explicit hydrogens and unique molecule ID\n\n### Step 3: Predict\
    \ BDE for Bonds\n**Single Bond**: **POST** `/api/v1/predict/single/` - Predict\
    \ BDE for one specific bond\n**Multiple Bonds**: **POST** `/api/v1/predict/multiple/`\
    \ - Predict BDE for multiple bonds\n**All Bonds**: **POST** `/api/v1/infer/all/`\
    \ - Predict BDE for all fragmentable bonds\n\n### Step 4: Fragment Analysis\n\
    **POST** `/api/v1/BDEEvaluate/` - Generate molecular fragments in SMILES or XYZ\
    \ format\n**POST** `/api/v1/ObtainBDEFragments/` - Find bond that generates specific\
    \ fragments\n**POST** `/api/v1/predict/check/` - Verify fragments generated by\
    \ bond cleavage\n\n### Step 5: Generate Reports\n**POST** `/api/v1/download_report/`\
...
paths:
...
  /api/v1/BDEEvaluate/:
    post:
      operationId: v1_BDEEvaluate_create
      description: "\n        Genera fragmentos moleculares en formato SMILES o XYZ.\n\
        \        Generates molecular fragments in SMILES or XYZ format.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BDEEvaluateRequest'
            examples:
              EntradaDeEjemplo/ExampleInput:
                value:
                  smiles: CCO
                  molecule_id: a1b2c3d4e5f6a7b8
                  bonds_idx:
                  - 0
                  - 1
                  export_smiles: true
                  export_xyz: false
                summary: Entrada de ejemplo / Example input
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BDEEvaluateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BDEEvaluateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[FragmentResponseData]'
              examples:
                EntradaDeEjemplo/ExampleInput:
                  value:
                    smiles: CCO
                    molecule_id: a1b2c3d4e5f6a7b8
                    bonds_idx:
                    - 0
                    - 1
                    export_smiles: true
                    export_xyz: false
                  summary: Entrada de ejemplo / Example input
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      smiles_canonical: CCO
                      molecule_id: a1b2c3d4e5f6a7b8
                      bonds_predicted: []
                      smiles_list: null
                      xyz_block: null
                      image_svg: <svg>...</svg>
                      canvas:
                        width: 300
                        height: 300
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorFormatoInválido:
                  value:
                    status: error
                    detail: Formato no soportado
                  summary: Error por formato inválido
          description: Error de validación / Validation error
  /api/v1/download_report/:
    post:
      operationId: v1_download_report_create
      description: "\n        Genera y descarga un informe txt con los resultados\
        \ de la predicción para la molécula \n        Generates and downloads a TXT\
        \ report with the prediction results for the indicated molecule \n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadReportRequest'
            examples:
              EntradaDeEjemplo/ExampleInput:
                value:
                  smiles: CCO
                  format: txt
                summary: Entrada de ejemplo / Example input
              EntradaAlternativa/AlternativeInput:
                value:
                  smiles: C1=CC=CC=C1
                  format: txt
                summary: Entrada alternativa / Alternative input
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DownloadReportRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DownloadReportRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[DownloadReportResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      type: pdf
                      report_base64: JVBERi0xLjQKJcfs...
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorFormatoInválido:
                  value:
                    status: error
                    detail: Formato no soportado
                  summary: Error por formato inválido
          description: Error de validación / Validation error
  /api/v1/infer/all/:
    post:
      operationId: v1_infer_all_create
      description: "\n        Predice las energías de disociación para todos los enlaces\
        \ simples de la molécula dada.\n        Predicts the dissociation energies\
        \ for all single bonds of the given molecule.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferAllRequest'
            examples:
              EntradaDeEjemplo/ExampleInput:
                value:
                  smiles: CCO
                summary: Entrada de ejemplo / Example input
              EntradaAlternativa/AlternativeInput:
                value:
                  smiles: C1=CC=CC=C1
                summary: Entrada alternativa / Alternative input
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InferAllRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InferAllRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[InferAllResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      smiles_canonical: CCO
                      molecule_id: a1b2c3d4e5f6a7b8
                      bonds:
                      - idx: 1
                        bde: 95.0
                        begin_atom_idx: 0
                        end_atom_idx: 1
                        bond_atoms: C-O
                        bond_type: single
                        is_fragmentable: true
                      - idx: 2
                        bde: 100.0
                        begin_atom_idx: 1
                        end_atom_idx: 2
                        bond_atoms: C-C
                        bond_type: single
                        is_fragmentable: true
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorSMILESInválido:
                  value:
                    status: error
                    detail: Formato SMILES inválido
                  summary: Error por SMILES inválido
          description: Error de validación / Validation error
  /api/v1/predict/check/:
    post:
      operationId: v1_predict_check_create
      description: "\n        Verifica productos generados por la escisión de un enlace\
        \ y predice la BDE.\n        Verifies products generated by the cleavage of\
        \ a bond and predicts the BDE.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictCheckRequest'
            examples:
              EntradaDeEjemplo/ExampleInput:
                value:
                  smiles: CCO
                  bond_idx: 1
                  molecule_id: 150018eccd174140
                  products:
                  - '[CH2][OH]'
                  - '[CH3]'
                summary: Entrada de ejemplo / Example input
              EntradaAlternativa/AlternativeInput:
                value:
                  smiles: CCO
                  bond_idx: 1
                  molecule_id: 150018eccd174140
                  products:
                  - '[CH2][OH]'
                  - '[CH3]'
                summary: Entrada alternativa / Alternative input
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PredictCheckRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PredictCheckRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[PredictCheckResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      smiles_canonical: CCO
                      bond:
                        idx: 1
                        bde: 95.0
                        begin_atom_idx: 0
                        end_atom_idx: 1
                        bond_atoms: C-O
                        bond_type: single
                        is_fragmentable: true
                      are_same_products: true
                      products:
                      - CC
                      - O
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorProductosInválidos:
                  value:
                    status: error
                    detail: Productos no coinciden con los esperados
                  summary: Error por productos inválidos
          description: Error de validación / Validation error
  /api/v1/predict/info/:
    post:
      operationId: v1_predict_info_create
      description: "\n        Devuelve información enriquecida de la molécula: SMILES\
        \ canónico, imagen SVG, metadatos de canvas, posiciones de átomos y enlaces,\
        \ id único.\n        \n        Returns enriched information about the molecule:\
        \ canonical SMILES, SVG image, canvas metadata, atom and bond positions, unique\
        \ ID.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoleculeInfoRequest'
            examples:
              EjemploDeEntrada/ExampleInput:
                value:
                  smiles: CCO
                summary: Ejemplo de entrada / Example input
              EjemploAlternativo/AlternativeExample:
                value:
                  smiles: C1=CC=CC=C1
                summary: Ejemplo alternativo / Alternative example
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MoleculeInfoRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MoleculeInfoRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[MoleculeInfoResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      smiles_canonical: CCO
                      image_svg: <svg>...</svg>
                      canvas:
                        width: 300
                        height: 300
                      atoms:
                        '0':
                          x: 50.0
                          y: 100.0
                          symbol: C
                          smiles: '[CH3]'
                        '1':
                          x: 150.0
                          y: 100.0
                          symbol: C
                          smiles: '[CH2]'
                        '2':
                          x: 250.0
                          y: 100.0
                          symbol: O
                          smiles: O
                      bonds:
                        '0':
                          start: 0
                          end: 1
                          start_coords:
                            x: 50.0
                            y: 100.0
                          end_coords:
                            x: 150.0
                            y: 100.0
                          bond_atoms: C-C
                          bond_type: single
                        '1':
                          start: 1
                          end: 2
                          start_coords:
                            x: 150.0
                            y: 100.0
                          end_coords:
                            x: 250.0
                            y: 100.0
                          bond_atoms: C-O
                          bond_type: single
                      molecule_id: a1b2c3d4e5f6a7b8
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorSMILESInválido:
                  value:
                    status: error
                    detail: Formato SMILES inválido
                  summary: Error por SMILES inválido
          description: Error de validación / Validation error
  /api/v1/predict/info-smile-canonical/:
    post:
      operationId: v1_predict_info_smile_canonical_create
      description: "        \n        Devuelve el SMILES canónico de una molécula\
        \ con hidrógenos explícitos.\n        Returns the canonical SMILES of a molecule\
        \ with explicit hydrogens.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoleculeSmileCanonicalRequest'
            examples:
              EjemploDeEntrada/ExampleInput:
                value:
                  smiles: CCO
                summary: Ejemplo de entrada / Example input
              EjemploAlternativo/AlternativeExample:
                value:
                  smiles: C1=CC=CC=C1
                summary: Ejemplo alternativo / Alternative example
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MoleculeSmileCanonicalRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MoleculeSmileCanonicalRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[MoleculeInfoResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data: CCO
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorSMILESInválido:
                  value:
                    status: error
                    detail: Formato SMILES inválido
                  summary: Error por SMILES inválido
          description: Error de validación / Validation error
  /api/v1/predict/multiple/:
    post:
      operationId: v1_predict_multiple_create
      description: "\n        Predice las energías de disociación para varios enlaces\
        \ de la molécula.\n        Predicts the dissociation energies for multiple\
        \ bonds of the molecule.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictMultipleRequest'
            examples:
              EntradaDeEjemplo/ExampleInput:
                value:
                  smiles: CCO
                  bond_indices:
                  - 1
                  - 2
                  molecule_id: 150018eccd174140
                summary: Entrada de ejemplo / Example input
              EntradaAlternativa/AlternativeInput:
                value:
                  smiles: C1=CC=CC=C1
                  bond_indices:
                  - 0
                  - 2
                  - 3
                  molecule_id: h1g2f3e4d5c6b7a8
                summary: Entrada alternativa / Alternative input
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PredictMultipleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PredictMultipleRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[PredictMultipleResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      smiles: CCO
                      molecule_id: a1b2c3d4e5f6a7b8
                      bonds:
                      - idx: 1
                        bde: 95.0
                        begin_atom_idx: 0
                        end_atom_idx: 1
                        bond_atoms: C-O
                        bond_type: single
                        is_fragmentable: true
                      - idx: 2
                        bde: 100.0
                        begin_atom_idx: 1
                        end_atom_idx: 2
                        bond_atoms: C-C
                        bond_type: single
                        is_fragmentable: true
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorÍndicesDeEnlaceInválidos:
                  value:
                    status: error
                    error:
                      code: BOND_INDEX_OUT_OF_RANGE
                      message: Índices de enlace fuera de rango
                  summary: Error por índices de enlace inválidos
          description: Error de validación / Validation error
  /api/v1/predict/single/:
    post:
      operationId: v1_predict_single_create
      description: "\n        Predice la energía de disociación para un enlace específico\
        \ de la molécula.\n        \n        Predicts the dissociation energy for\
        \ a specific bond of the molecule.\n        "
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictSingleRequest'
            examples:
              EjemploDeEntrada/ExampleInput:
                value:
                  smiles: CCO
                  bond_idx: 1
                  molecule_id: 150018eccd174140
                summary: Ejemplo de entrada / Example input
              EjemploAlternativo/AlternativeExample:
                value:
                  smiles: C1=CC=CC=C1
                  bond_idx: 2
                  molecule_id: 4b7620ed22c55dfd
                summary: Ejemplo alternativo / Alternative example
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PredictSingleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PredictSingleRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[PredictSingleResponseData]'
              examples:
                EjemploExitoso:
                  value:
                    status: success
                    data:
                      smiles_canonical: CCO
                      bond:
                        idx: 1
                        bde: 95.0
                        begin_atom_idx: 0
                        end_atom_idx: 1
                        bond_atoms: C-O
                        bond_type: single
                        is_fragmentable: true
                    error: null
                  summary: Ejemplo exitoso
          description: Respuesta exitosa / Successful response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse[NoneType]'
              examples:
                ErrorPorÍndiceDeEnlaceInválido:
                  value:
                    status: error
                    detail: Índice de enlace fuera de rango
                  summary: Error por índice de enlace inválido
          description: Error de validación / Validation error
components:
  schemas:
...
    BDEEvaluateRequest:
      description: |-
        Entrada para /fragment/.
        - smiles: SMILES con hidrógenos explícitos
        - molecule_id: ID de la molécula (opcional si se proporcionan fragmentos)
        - bonds_idx: Lista de índices de enlaces (opcional si se proporcionan fragmentos)
        - export_smiles: Indica si se debe exportar la lista de SMILES de los fragmentos
        - export_xyz: Indica si se debe exportar el bloque XYZ de los fragmentos
      properties:
        smiles:
          title: Smiles
          type: string
        molecule_id:
          title: Molecule Id
          type: string
        bonds_idx:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Bonds Idx
        export_smiles:
          default: false
          title: Export Smiles
          type: boolean
        export_xyz:
          default: false
          title: Export Xyz
          type: boolean
      required:
      - smiles
      - molecule_id
      - bonds_idx
      title: BDEEvaluateRequest
      type: object
...
    InferAllRequest:
      description: |-
        Entrada para /infer/all/.
        - smiles: SMILES de la molécula
      properties:
        smiles:
          title: Smiles
          type: string
      required:
      - smiles
      title: InferAllRequest
      type: object
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid